refactor(jwt): address review feedback on PR #20513
- Extract test helpers into dedicated util files (decode-jwt-complete-or-throw, forge-legacy-hs256-token, renew-token, generate-api-key-token, renew-application-token)
- Split jwt-key-rotation integration spec into two suites: jwt-asymmetric-signing (new ES256 + kid) and jwt-legacy-hs256-fallback (legacy HS256)
- Drop the it.each block that was effectively testing beforeAll output
- Rename JwtWrapperService.signAsync -> signAsyncOrThrow to make the throw contract explicit; update all callers and unit tests
- Add missing WORKSPACE_AGNOSTIC token-type guard in WorkspaceAgnosticTokenService.validateToken (matches login-token pattern)
- Parallelize availableWorkspacesForSignUp/SignIn in setLoginTokenToAvailableWorkspacesWhenAuthProviderMatch
