+++ packages/twenty-server/src/engine/core-modules/api-key/services/api-key.service.ts
@@ -157,7 +157,7 @@ export class ApiKeyService {
       expiresIn = '100y';
     }
 
-    const token = await this.jwtWrapperService.signAsync(
+    const token = await this.jwtWrapperService.signAsyncOrThrow(
       {
         sub: workspaceId,
         type: JwtTokenTypeEnum.API_KEY,
