+++ packages/twenty-server/src/engine/core-modules/auth/token/services/access-token.service.ts
@@ -141,7 +141,7 @@ export class AccessTokenService {
       impersonatedUserWorkspaceId: payloadOriginalUserWorkspaceId,
     };
 
-    const token = await this.jwtWrapperService.signAsync(jwtPayload, {
+    const token = await this.jwtWrapperService.signAsyncOrThrow(jwtPayload, {
       expiresIn,
     });
