+++ packages/twenty-server/src/engine/core-modules/auth/token/services/application-token.service.spec.ts
@@ -29,7 +29,7 @@ describe('ApplicationTokenService', () => {
         {
           provide: JwtWrapperService,
           useValue: {
-            signAsync: jest.fn(),
+            signAsyncOrThrow: jest.fn(),
             verifyJwtToken: jest.fn(),
             decode: jest.fn(),
             extractJwtFromRequest: jest.fn(),
@@ -80,7 +80,9 @@ describe('ApplicationTokenService', () => {
       jest
         .spyOn(applicationRepository, 'findOne')
         .mockResolvedValue(mockApplication as ApplicationEntity);
-      jest.spyOn(jwtWrapperService, 'signAsync').mockResolvedValue(mockToken);
+      jest
+        .spyOn(jwtWrapperService, 'signAsyncOrThrow')
+        .mockResolvedValue(mockToken);
 
       const result = await service.generateApplicationAccessToken({
         workspaceId,
@@ -91,7 +93,7 @@ describe('ApplicationTokenService', () => {
         token: mockToken,
         expiresAt: expect.any(Date),
       });
-      expect(jwtWrapperService.signAsync).toHaveBeenCalledWith(
+      expect(jwtWrapperService.signAsyncOrThrow).toHaveBeenCalledWith(
         expect.objectContaining({
           sub: applicationId,
           applicationId,
@@ -115,7 +117,9 @@ describe('ApplicationTokenService', () => {
       jest
         .spyOn(applicationRepository, 'findOne')
         .mockResolvedValue(mockApplication as ApplicationEntity);
-      jest.spyOn(jwtWrapperService, 'signAsync').mockResolvedValue(mockToken);
+      jest
+        .spyOn(jwtWrapperService, 'signAsyncOrThrow')
+        .mockResolvedValue(mockToken);
 
       const result = await service.generateApplicationAccessToken({
         workspaceId,
@@ -128,7 +132,7 @@ describe('ApplicationTokenService', () => {
         token: mockToken,
         expiresAt: expect.any(Date),
       });
-      expect(jwtWrapperService.signAsync).toHaveBeenCalledWith(
+      expect(jwtWrapperService.signAsyncOrThrow).toHaveBeenCalledWith(
         expect.objectContaining({
           sub: applicationId,
           applicationId,
@@ -270,7 +274,9 @@ describe('ApplicationTokenService', () => {
       jest
         .spyOn(applicationRepository, 'findOne')
         .mockResolvedValue(mockApplication as ApplicationEntity);
-      jest.spyOn(jwtWrapperService, 'signAsync').mockResolvedValue(mockToken);
+      jest
+        .spyOn(jwtWrapperService, 'signAsyncOrThrow')
+        .mockResolvedValue(mockToken);
 
       const result = await service.generateApplicationTokenPair({
         workspaceId,
@@ -285,7 +291,7 @@ describe('ApplicationTokenService', () => {
         token: mockToken,
         expiresAt: expect.any(Date),
       });
-      expect(jwtWrapperService.signAsync).toHaveBeenCalledTimes(2);
+      expect(jwtWrapperService.signAsyncOrThrow).toHaveBeenCalledTimes(2);
     });
   });
 
@@ -303,7 +309,9 @@ describe('ApplicationTokenService', () => {
       jest
         .spyOn(applicationRepository, 'findOne')
         .mockResolvedValue(mockApplication as ApplicationEntity);
-      jest.spyOn(jwtWrapperService, 'signAsync').mockResolvedValue(mockToken);
+      jest
+        .spyOn(jwtWrapperService, 'signAsyncOrThrow')
+        .mockResolvedValue(mockToken);
 
       const result = await service.renewApplicationTokens({
         workspaceId,
