+++ packages/twenty-server/src/engine/core-modules/auth/token/services/login-token.service.ts
@@ -42,7 +42,7 @@ export class LoginTokenService {
     const expiresAt = addMilliseconds(new Date().getTime(), ms(expiresIn));
 
     return {
-      token: await this.jwtWrapperService.signAsync(jwtPayload, {
+      token: await this.jwtWrapperService.signAsyncOrThrow(jwtPayload, {
         expiresIn,
       }),
       expiresAt,
