+++ packages/twenty-server/src/engine/core-modules/auth/token/services/refresh-token.service.spec.ts
@@ -31,7 +31,7 @@ describe('RefreshTokenService', () => {
           useValue: {
             verifyJwtToken: jest.fn(),
             decode: jest.fn(),
-            signAsync: jest.fn(),
+            signAsyncOrThrow: jest.fn(),
             generateAppSecret: jest.fn(),
           },
         },
@@ -125,7 +125,9 @@ describe('RefreshTokenService', () => {
       const mockExpiresIn = '7d';
 
       jest.spyOn(twentyConfigService, 'get').mockReturnValue(mockExpiresIn);
-      jest.spyOn(jwtWrapperService, 'signAsync').mockResolvedValue(mockToken);
+      jest
+        .spyOn(jwtWrapperService, 'signAsyncOrThrow')
+        .mockResolvedValue(mockToken);
       jest
         .spyOn(appTokenRepository, 'create')
         .mockReturnValue({ id: 'new-token-id' } as AppTokenEntity);
@@ -144,7 +146,7 @@ describe('RefreshTokenService', () => {
         expiresAt: expect.any(Date),
       });
       expect(appTokenRepository.save).toHaveBeenCalled();
-      expect(jwtWrapperService.signAsync).toHaveBeenCalledWith(
+      expect(jwtWrapperService.signAsyncOrThrow).toHaveBeenCalledWith(
         {
           sub: userId,
           workspaceId,
