+++ packages/twenty-server/src/engine/core-modules/file/file-url/file-url.service.ts
@@ -52,7 +52,7 @@ export class FileUrlService {
       type: JwtTokenTypeEnum.FILE,
     };
 
-    const token = await this.jwtWrapperService.signAsync(payload, {
+    const token = await this.jwtWrapperService.signAsyncOrThrow(payload, {
       expiresIn: fileTokenExpiresIn,
     });
